# server/docs/openapi.yaml
openapi: "3.2.0"
info:
  title: Language App API
  version: "1.0.0"
  description: |
    REST API for the Language App. Segments Chinese text into words,
    provides pinyin transliteration, and English translations.
    Includes OCR text extraction, SRS vocabulary review, and segment editing.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - sessionCookie: []
...
paths:
...
  /api/texts:
    post:
      tags: [texts]
      summary: Create a text record
      operationId: createText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [raw_text, source_type]
              properties:
                raw_text:
                  type: string
                source_type:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Text created
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/texts/{text_id}:
    get:
      tags: [texts]
      summary: Get a text record
      operationId: getText
      parameters:
        - name: text_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Text record
          content:
            application/json:
              schema:
                type: object
                required: [id, created_at, source_type, raw_text, normalized_text]
                properties:
                  id:
                    type: string
                  created_at:
                    type: string
                    format: date-time
                  source_type:
                    type: string
                  raw_text:
                    type: string
                  normalized_text:
                    type: string
                  metadata:
                    type: object
                    additionalProperties: true
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/events:
    post:
      tags: [events]
      summary: Create an interaction event
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_type]
              properties:
                event_type:
                  type: string
                text_id:
                  type: ["string", "null"]
                segment_id:
                  type: ["string", "null"]
                payload:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Event created
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/vocab/save:
    post:
      tags: [vocab]
      summary: Save a vocabulary item
      operationId: saveVocab
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [headword, pinyin, english, status]
              properties:
                headword:
                  type: string
                pinyin:
                  type: string
                english:
                  type: string
                text_id:
                  type: ["string", "null"]
                segment_id:
                  type: ["string", "null"]
                snippet:
                  type: ["string", "null"]
                status:
                  type: string
      responses:
        "200":
          description: Vocab item saved
          content:
            application/json:
              schema:
                type: object
                required: [vocab_item_id]
                properties:
                  vocab_item_id:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/vocab/status:
    post:
      tags: [vocab]
      summary: Update vocabulary item status
      operationId: updateVocabStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vocab_item_id, status]
              properties:
                vocab_item_id:
                  type: string
                status:
                  type: string
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/vocab/lookup:
    post:
      tags: [vocab]
      summary: Record a vocabulary lookup
      operationId: recordLookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vocab_item_id]
              properties:
                vocab_item_id:
                  type: string
      responses:
        "200":
          description: Lookup recorded with SRS state
          content:
            application/json:
              schema:
                type: object
                required: [vocab_item_id, opacity, is_struggling]
                properties:
                  vocab_item_id:
                    type: string
                  opacity:
                    type: number
                    format: float
                    description: "1.0 = new/struggling, 0.0 = known"
                  is_struggling:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/vocab/srs-info:
    get:
      tags: [vocab]
      summary: Get SRS info for headwords
      operationId: getVocabSRSInfo
      parameters:
        - name: headwords
          in: query
          description: Comma-separated list of headwords
          schema:
            type: string
      responses:
        "200":
          description: SRS info for matching vocab items
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/VocabSRSItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/review/words/queue:
    get:
      tags: [review]
      summary: Get SRS review queue
      operationId: getReviewQueue
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Review cards due
          content:
            application/json:
              schema:
                type: object
                required: [cards, due_count]
                properties:
                  cards:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReviewCard"
                  due_count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/review/answer:
    post:
      tags: [review]
      summary: Record a review answer
      operationId: recordReviewAnswer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vocab_item_id, grade]
              properties:
                vocab_item_id:
                  type: string
                grade:
                  type: integer
      responses:
        "200":
          description: Review answer recorded
          content:
            application/json:
              schema:
                type: object
                required: [vocab_item_id, interval_days, remaining_due]
                properties:
                  vocab_item_id:
                    type: string
                  next_due_at:
                    type: ["string", "null"]
                    format: date-time
                  interval_days:
                    type: number
                    format: float
                  remaining_due:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/review/words/count:
    get:
      tags: [review]
      summary: Get count of word reviews due
      operationId: getReviewCount
      responses:
        "200":
          description: Count of due word reviews
          content:
            application/json:
              schema:
                type: object
                required: [due_count]
                properties:
                  due_count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/review/characters/queue:
    get:
      tags: [review]
      summary: Get character SRS review queue
      operationId: getCharacterReviewQueue
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Character review cards due
          content:
            application/json:
              schema:
                type: object
                required: [cards, due_count]
                properties:
                  cards:
                    type: array
                    items:
                      $ref: "#/components/schemas/CharacterReviewCard"
                  due_count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/review/characters/count:
    get:
      tags: [review]
      summary: Get count of character reviews due
      operationId: getCharacterReviewCount
      responses:
        "200":
          description: Count of due character reviews
          content:
            application/json:
              schema:
                type: object
                required: [due_count]
                properties:
                  due_count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/segments/translate-batch:
    post:
      tags: [segments]
      summary: Translate a batch of segments
      operationId: translateBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [segments]
              properties:
                segments:
                  type: array
                  items:
                    type: string
                context:
                  type: ["string", "null"]
                translation_id:
                  type: ["string", "null"]
                paragraph_idx:
                  type: ["integer", "null"]
      responses:
        "200":
          description: Translated segments
          content:
            application/json:
              schema:
                type: object
                required: [translations]
                properties:
                  translations:
                    type: array
                    items:
                      $ref: "#/components/schemas/SegmentTranslation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...
    SegmentTranslation:
      type: object
      required: [segment, pinyin, english]
      properties:
        segment:
          type: string
        pinyin:
          type: string
        english:
          type: string
    VocabSRSItem:
      type: object
      required: [vocab_item_id, headword, pinyin, english, opacity, is_struggling, status]
      properties:
        vocab_item_id:
          type: string
        headword:
          type: string
        pinyin:
          type: string
        english:
          type: string
        opacity:
          type: number
          format: float
          description: "1.0 = new/struggling, 0.0 = known"
        is_struggling:
          type: boolean
        status:
          type: string
    ReviewCard:
      type: object
      required: [vocab_item_id, headword, pinyin, english, snippets]
      properties:
        vocab_item_id:
          type: string
        headword:
          type: string
        pinyin:
          type: string
        english:
          type: string
        snippets:
          type: array
          items:
            type: string
...
    CharacterReviewCard:
      type: object
      required: [vocab_item_id, character, pinyin, english, example_words]
      properties:
        vocab_item_id:
          type: string
        character:
          type: string
        pinyin:
          type: string
        english:
          type: string
        example_words:
          type: array
          items:
            $ref: "#/components/schemas/CharacterExampleWord"
    CharacterExampleWord:
      type: object
      required: [vocab_item_id, headword, pinyin, english]
      properties:
        vocab_item_id:
          type: string
        headword:
          type: string
        pinyin:
          type: string
        english:
          type: string